Added SQL connection to page
Just added a quick for loop so that the page actually looks for things from the sql, and view details button brings you to the actual item page. I will be working on getting that functionality working, while also adding more info. We will need to add a description column to products.